requirements.txt
Django
mysqlclient
django-environ==0.4.5
django-cities-light
django-bootstrap4
django-bootstrap-datepicker-plus
djangorestframework==3.14.0